Allow plugins to be hidden from the plugin list
A plugin developer can choose to have their plugin hidden by default from the plugin list. This is useful if devs/servers who want to deploy small plugins and do not want to clutter their plugin list or to hide extension plugins. This can be enabled by placing a tag in the plugin.yml file called `visible: true` to hide the plugin when using Project Poseidon. Plugins that are disabled will always appear in the plugin list so errors can be easily identified.