Implement pom.xml for building with Maven (#26)
Uses the pom similar to what CraftBukkit used at the commit of the fork.
Some changes needed to be made due to dependencies either not existing or being out of date.
Bukkit needed to be compiled, it was compiled as of the commit listed in the Readme and is hosted on my build server.

com.sun.istack.internals.NotNull had to be replaced with com.jetbrains.annotations.NotNull. Should not be an issue.