Fix dupe glitches/strategies and stop protection workarounds.
Update Warning: Critical
This update removes the Bukkit chunk cache which as a result of underlying code issues can be used to dupe, exploit, and get around plugin-based protections due to the incorrect reporting of block information. The chunk cache was removed from contemporary Bukkit in 2012, assumably for the same reason.